Fixed Neuron Code

+++ src/math/tensor.rs
@@ -3,5 +3,5 @@
 #[derive(Clone)]
 pub struct Tensor {
     shape: Vec<usize>,
-    data: Vec<f32>,
+    data: Vec<f64>,
 }
